Remove project view
